add composer script
