+++ composer.json
@@ -47,9 +47,12 @@
         "php-di/php-di": "^6.0"
     },
     "require-dev": {
-        "phpunit/phpunit": ">=8.3",
+        "phpunit/phpunit": "^8",
         "phpdocumentor/phpdocumentor": "2.*",
         "evert/phpdoc-md": "0.2.0",
         "phly/changelog-generator": "^2.1"
-    }
+    },
+    "scripts": {
+        "test:unit": "./vendor/phpunit/phpunit/phpunit -c ./phpunit.xml --testsuite=unit --verbose"
+      }
 }
