+++ phpunit.xml
@@ -7,8 +7,7 @@
          convertNoticesToExceptions="true"
          convertWarningsToExceptions="true"
          processIsolation="false"
-         stopOnFailure="false"
-         syntaxCheck="false">
+         stopOnFailure="false">
     <testsuites>
         <testsuite name="Backend Proxy Test Suite">
             <directory suffix=".php">./tests/</directory>
@@ -24,6 +23,6 @@
     </filter>
     <logging>
         <log type="coverage-clover" target="test-results/coverage.xml"/>
-        <log type="junit" target="test-results/junit.xml" logIncompleteSkipped="false"/>
+        <log type="junit" target="test-results/junit.xml"/>
     </logging>
 </phpunit>
