+++ tests/unit/CookieProviderTest.php
