+++ tests/unit/DeviceCodeControllerTest.php
@@ -6,16 +6,17 @@ use GuzzleHttp\Psr7\Response;
 
 use PHPUnit\Framework\TestCase;
 
-use NPR\One\Controllers\AbstractOAuth2Controller;
+use NPR\One\Controllers\DeviceCodeController;
 use NPR\One\DI\DI;
 use NPR\One\Interfaces\ConfigInterface;
-use NPR\One\Models\AccessTokenModel;
+use NPR\One\Models\{AccessTokenModel, DeviceCodeModel};
 use NPR\One\Providers\{CookieProvider, EncryptionProvider, SecureCookieProvider};
 
-class CustomControllerTests extends TestCase
+class DeviceCodeControllerTests extends TestCase
 {
     const ACCESS_TOKEN_RESPONSE = '{"access_token": "LT8gvVDyeKwQJVVf6xwKAWdK0bOik64faketoken","token_type": "Bearer","expires_in": 690448786,"refresh_token": "6KVn9BOhHhUFR1Yqi2T2pzpTWI9WIfakerefresh"}';
     const ACCESS_TOKEN_RESPONSE_2 = '{"access_token": "LT8gvVDyeKwQJVVf6xwKAWdK0bOik64faketoken","token_type": "Bearer","expires_in": 690448786}';
+    const DEVICE_CODE_RESPONSE = '{"device_code":"IevXEi6eNBPemJA7OWCuBzQ3tua9iHyifakecode","user_code":"2OA7PP","verification_uri":"http:\/\/www.npr.org\/device","expires_in":1800,"interval":5}';
 
     /** @var SecureCookieProvider */
     private $mockSecureCookie;
@@ -30,7 +31,7 @@ class CustomControllerTests extends TestCase
     private static $clientId = 'fake_client_id';
 
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->mockSecureCookie = $this->getMock(SecureCookieProvider::class);
 
@@ -57,80 +58,62 @@ class CustomControllerTests extends TestCase
      */
     public function testConfigProviderException()
     {
-        $controller = new CustomController();
-        $controller->issueAccessToken();
-    }
-
-    /**
-     * @expectedException \RuntimeException
-     * @expectedExceptionMessageRegExp   #SecureStorageProvider must be set. See.*setSecureStorageProvider#
-     */
-    public function testSecureStorageProviderException()
-    {
-        $controller = new CustomController();
-        $controller->setConfigProvider($this->mockConfig);
-        $controller->issueAccessToken();
+        $controller = new DeviceCodeController();
+        $controller->startDeviceCodeGrant(['fake_scope']);
     }
 
     /**
      * @expectedException \RuntimeException
      * @expectedExceptionMessageRegExp   #WARNING: It is strongly discouraged to use CookieProvider as your secure storage provider.#
      */
-    public function testSecureStorageProviderWarning()
+    public function testSecureStorageProviderException()
     {
         $mockCookie = $this->getMock(CookieProvider::class);
 
-        $controller = new CustomController();
+        $controller = new DeviceCodeController();
         $controller->setConfigProvider($this->mockConfig);
         $controller->setSecureStorageProvider($mockCookie);
-        $controller->issueAccessToken();
-    }
-
-    /**
-     * @expectedException \RuntimeException
-     * @expectedExceptionMessageRegExp   #EncryptionProvider must be set. See.*setEncryptionProvider#
-     */
-    public function testEncryptionProviderException()
-    {
-        $controller = new CustomController();
-        $controller->setConfigProvider($this->mockConfig);
-        $controller->setSecureStorageProvider($this->mockSecureCookie);
-        $controller->issueAccessToken();
+        $controller->startDeviceCodeGrant(['fake_scope']);
     }
 
     /**
      * @expectedException \RuntimeException
      * @expectedExceptionMessageRegExp   #EncryptionProvider must be valid. See.*EncryptionInterface::isValid#
      */
-    public function testEncryptionProviderInvalidException()
+    public function testEncryptionProviderException()
     {
         $mockEncryption = $this->getMock(EncryptionProvider::class);
         $mockEncryption->method('isValid')->willReturn(false);
 
-        $controller = new CustomController();
+        $controller = new DeviceCodeController();
         $controller->setConfigProvider($this->mockConfig);
-        $controller->setSecureStorageProvider($this->mockSecureCookie);
         $controller->setEncryptionProvider($mockEncryption);
-        $controller->issueAccessToken();
+        $controller->startDeviceCodeGrant(['fake_scope']);
     }
 
-    /**
-     * @expectedException \InvalidArgumentException
-     * @expectedExceptionMessage   Must specify grant type
-     */
-    public function testBadIssueAccessTokenMissingGrantType()
+    public function testStartDeviceCodeGrant()
     {
-        $controller = new CustomController();
+        $mock = new MockHandler([
+            new Response(200, [], self::DEVICE_CODE_RESPONSE),
+        ]);
+
+        $handler = HandlerStack::create($mock);
+        $client = new Client(['handler' => $handler]);
+
+        DI::container()->set(Client::class, $client);
+
+        $controller = new DeviceCodeController();
         $controller->setConfigProvider($this->mockConfig);
-        $controller->setSecureStorageProvider($this->mockSecureCookie);
-        $controller->setEncryptionProvider($this->mockEncryption);
-        $controller->badIssueAccessToken();
+        $deviceCode = $controller->startDeviceCodeGrant(['fake_scope']);
+
+        $this->assertInstanceOf(DeviceCodeModel::class, $deviceCode, 'startDeviceCodeGrant response was not of type DeviceCodeModel: ' . print_r($deviceCode, 1));
+        $this->assertEquals(0, $mock->count(), 'Expected additional HTTP requests to be made');
     }
 
     /**
      * @expectedException \Exception
      */
-    public function testIssueAccessTokenWithApiError()
+    public function testStartDeviceCodeGrantWithApiException()
     {
         $mock = new MockHandler([
             new Response(500, [], ''),
@@ -138,14 +121,26 @@ class CustomControllerTests extends TestCase
 
         $handler = HandlerStack::create($mock);
         $client = new Client(['handler' => $handler]);
+
         DI::container()->set(Client::class, $client);
 
-        $controller = new CustomController();
+        $controller = new DeviceCodeController();
         $controller->setConfigProvider($this->mockConfig);
-        $controller->issueAccessToken();
+        $controller->startDeviceCodeGrant(['fake_scope']);
     }
 
-    public function testIssueAccessToken()
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage   Could not locate a device code
+     */
+    public function testPollDeviceCodeGrantMissingDeviceCode()
+    {
+        $controller = new DeviceCodeController();
+        $controller->setConfigProvider($this->mockConfig);
+        $controller->pollDeviceCodeGrant();
+    }
+
+    public function testPollDeviceCodeGrant()
     {
         $mock = new MockHandler([
             new Response(200, [], self::ACCESS_TOKEN_RESPONSE),
@@ -153,19 +148,20 @@ class CustomControllerTests extends TestCase
 
         $handler = HandlerStack::create($mock);
         $client = new Client(['handler' => $handler]);
+
         DI::container()->set(Client::class, $client);
 
-        $controller = new CustomController();
-        $controller->setConfigProvider($this->mockConfig);
-        $controller->setSecureStorageProvider($this->mockSecureCookie);
-        $controller->setEncryptionProvider($this->mockEncryption);
-        $accessToken = $controller->issueAccessToken();
+        $this->mockSecureCookie->method('get')->willReturn('i_am_a_device_code');
 
-        $this->assertInstanceOf(AccessTokenModel::class, $accessToken, 'issueAccessToken response was not of type AccessTokenModel: ' . print_r($accessToken, 1));
+        $controller = new DeviceCodeController();
+        $controller->setConfigProvider($this->mockConfig);
+        $accessToken = $controller->pollDeviceCodeGrant();
+
+        $this->assertInstanceOf(AccessTokenModel::class, $accessToken, 'pollDeviceCodeGrant response was not of type AccessTokenModel: ' . print_r($accessToken, 1));
         $this->assertEquals(0, $mock->count(), 'Expected additional HTTP requests to be made');
     }
 
-    public function testIssueAccessTokenNoRefreshToken()
+    public function testPollDeviceCodeGrantNoRefreshToken()
     {
         $mock = new MockHandler([
             new Response(200, [], self::ACCESS_TOKEN_RESPONSE_2),
@@ -173,45 +169,16 @@ class CustomControllerTests extends TestCase
 
         $handler = HandlerStack::create($mock);
         $client = new Client(['handler' => $handler]);
+
         DI::container()->set(Client::class, $client);
 
-        $controller = new CustomController();
+        $this->mockSecureCookie->method('get')->willReturn('i_am_a_device_code');
+
+        $controller = new DeviceCodeController();
         $controller->setConfigProvider($this->mockConfig);
-        $controller->setSecureStorageProvider($this->mockSecureCookie);
-        $controller->setEncryptionProvider($this->mockEncryption);
-        $accessToken = $controller->issueAccessToken();
+        $accessToken = $controller->pollDeviceCodeGrant();
 
-        $this->assertInstanceOf(AccessTokenModel::class, $accessToken, 'issueAccessToken response was not of type AccessTokenModel: ' . print_r($accessToken, 1));
+        $this->assertInstanceOf(AccessTokenModel::class, $accessToken, 'pollDeviceCodeGrant response was not of type AccessTokenModel: ' . print_r($accessToken, 1));
         $this->assertEquals(0, $mock->count(), 'Expected additional HTTP requests to be made');
-    }
-}
-
-class CustomController extends AbstractOAuth2Controller
-{
-    public function __construct()
-    {
-        // intentionally don't call parent constructor
-    }
-
-    public function issueAccessToken()
-    {
-        $this->ensureExternalProvidersExist();
-
-        $accessToken = $this->createAccessToken('fake_grant_type');
-
-        $this->storeRefreshToken($accessToken);
-
-        return $accessToken;
-    }
-
-    public function badIssueAccessToken()
-    {
-        $this->ensureExternalProvidersExist();
-
-        $accessToken = $this->createAccessToken(null);
-
-        $this->storeRefreshToken($accessToken);
-
-        return $accessToken;
     }
 }
