+++ tests/unit/DeviceCodeModelTest.php
