+++ tests/unit/EncryptionProviderTest.php
